+++ src/com/runrunfast/homegym/record/StatisticalData.java
@@ -2,5 +2,5 @@ package com.runrunfast.homegym.record;
 
 public class StatisticalData {
 	public String strDate;
-	public int totalKcal;
+	public float totalKcal;
 }
