+++ src/com/runrunfast/homegym/record/TotalRecord.java
@@ -2,7 +2,7 @@ package com.runrunfast.homegym.record;
 
 public class TotalRecord {
 	public String uid;
-	public int total_kcal;
+	public float total_kcal;
 	public int total_time;
 	public int total_count;
 	public int total_train_count;
